+++ src/services/marketplace/index.ts
@@ -1,3 +1,4 @@
 export * from "./SimpleInstaller"
 export * from "./MarketplaceManager"
+export * from "./SkillsMarketplaceManager"
 export type { MarketplaceItemType } from "@roo-code/types"